+++ lib/naver/sdk.rb
@@ -2,6 +2,7 @@ require "oauth2"
 require "faraday"
 
 require "naver/sdk/version"
+require "naver/ext/hash/to_underscore_keys"
 require "naver/config"
 require "naver/connection"
 require "naver/client"